requirements.txt
_libgcc_mutex==0.1
blas==1.0
ca-certificates==2021.1.19
certifi==2020.12.5
cudatoolkit==10.0.130
cython==0.29.21
freetype==2.10.4
intel-openmp==2020.2
jpeg==9b
lcms2==2.11
libedit==3.1.20191231
libffi==3.2.1
libgcc-ng==9.1.0
libpng==1.6.37
libstdcxx-ng==9.1.0
libtiff==4.1.0
lz4-c==1.9.3
maskrcnn-benchmark==0.1
mkl==2020.2
mkl-service==2.3.0
mkl_fft==1.2.0
mkl_random==1.1.1
ncurses==6.2
ninja==1.10.0.post2
numpy==1.19.2
numpy-base==1.19.2
olefile==0.46
opencv-python==4.4.0.40
openssl==1.1.1i
pillow==7.2.0
pip==20.3.3
python==3.6.7
pytorch==1.4.0
pyyaml==5.3.1
readline==7.0
scipy==1.5.4
setuptools==52.0.0
six==1.15.0
sqlite==3.33.0
timm==0.1.30
tk==8.6.10
torchvision==0.5.0
tqdm==4.48.2
wheel==0.36.2
xz==5.2.5
yacs==0.1.8
zlib==1.2.11
zstd==1.4.5
# added by kimkk for pycharm
torch==1.4.0
torchinfo==1.4.4
onnx==1.8.0
